[feat] CalendarViewModel 로직을 MainViewModel로 옮김
