+++ app/src/main/java/com/eatssu/android/presentation/main/cafeteria/CafeteriaFragment.kt
@@ -6,12 +6,14 @@ import android.view.LayoutInflater
 import android.view.View
 import android.widget.TextView
 import androidx.annotation.RequiresApi
+import androidx.fragment.app.activityViewModels
 import androidx.lifecycle.ViewModelProvider
 import androidx.recyclerview.widget.GridLayoutManager
 import androidx.recyclerview.widget.RecyclerView
 import androidx.viewpager2.widget.ViewPager2
 import com.eatssu.android.databinding.FragmentCafeteriaBinding
 import com.eatssu.android.presentation.base.BaseFragment
+import com.eatssu.android.presentation.main.MainViewModel
 import com.eatssu.android.presentation.main.calendar.CalendarAdapter
 import com.eatssu.android.presentation.main.calendar.CalendarAdapter.OnItemListener
 import com.eatssu.android.presentation.main.calendar.CalendarViewModel
@@ -26,7 +28,9 @@ import java.time.LocalDate
 @AndroidEntryPoint
 class CafeteriaFragment : BaseFragment<FragmentCafeteriaBinding>(), OnItemListener {
 
-    private lateinit var calendarViewModel: CalendarViewModel
+//    private lateinit var calendarViewModel: CalendarViewModel
+
+    private val mainViewModel by activityViewModels<MainViewModel>()
 
     private var monthYearText: TextView? = null
     private var calendarRecyclerView: RecyclerView? = null
@@ -52,14 +56,13 @@ class CafeteriaFragment : BaseFragment<FragmentCafeteriaBinding>(), OnItemListen
 
         initWidgets()
         CalendarUtil.selectedDate = LocalDate.now()
-        calendarViewModel.setData(CalendarUtil.selectedDate)
+        mainViewModel.setData(CalendarUtil.selectedDate)
         setWeekView()
     }
 
     private fun initWidgets() {
         calendarRecyclerView = binding.weekRecycler
         monthYearText = binding.monthYearTV
-        calendarViewModel = ViewModelProvider(this)[CalendarViewModel::class.java]
     }
 
     @RequiresApi(Build.VERSION_CODES.O)
@@ -90,7 +93,7 @@ class CafeteriaFragment : BaseFragment<FragmentCafeteriaBinding>(), OnItemListen
     @RequiresApi(Build.VERSION_CODES.O)
     override fun onItemClick(position: Int, date: LocalDate) {
         CalendarUtil.selectedDate = date
-        calendarViewModel.setData(date)
+        mainViewModel.setData(date)
         mainPosition = position
         setWeekView()
     }
